@@ -1,39 +1,39 @@
 <?php
-// For php-cs-fixer 3.17.0
+// For php-cs-fixer 3.63.1
+// https://github.com/PHP-CS-Fixer/PHP-CS-Fixer#usage
+// https://mlocati.github.io/php-cs-fixer-configurator/#version:3.63
 return (new PhpCsFixer\Config())
     ->setRules([
-        '@PSR2'                           => true,
-        'array_indentation'               => true,
-        'array_syntax'                    => [
+        '@PSR2'             => true,
+        'array_indentation' => true,
+        'array_syntax'      => [
             'syntax' => 'short'
         ],
-        'combine_consecutive_unsets'      => true,
-        'binary_operator_spaces'          => [
+        'combine_consecutive_unsets' => true,
+        'binary_operator_spaces'     => [
             'operators' => [
-                '='  => 'align_single_space',
-                '=>' => 'align_single_space',
+                '='  => 'align_single_space_minimal',
+                '=>' => 'align_single_space_minimal',
             ]
         ],
-        'function_typehint_space'         => true,
-        'class_attributes_separation'     => [
+        'class_attributes_separation' => [
             'elements' => [
                 'method' => 'one'
             ]
         ],
-        'braces'                          => [
-            'allow_single_line_anonymous_class_with_empty_body' => true,
-            'allow_single_line_closure'                         => true,
+        'braces_position' => [
+            'allow_single_line_empty_anonymous_classes' => true,
         ],
-        'no_unused_imports'               => true,
-        'ordered_imports'                 => true,
-        'return_type_declaration'         => [
+        'no_unused_imports'       => true,
+        'ordered_imports'         => true,
+        'return_type_declaration' => [
             'space_before' => 'one'
         ],
-        'whitespace_after_comma_in_array' => true,
-        'no_superfluous_elseif'           => true,
-        'no_useless_else'                 => true,
-        'no_whitespace_in_blank_line'     => true,
-        'compact_nullable_typehint'       => true,
+        'whitespace_after_comma_in_array'   => true,
+        'no_superfluous_elseif'             => true,
+        'no_useless_else'                   => true,
+        'no_whitespace_in_blank_line'       => true,
+        'compact_nullable_type_declaration' => true,
     ])
     ->setLineEnding("\n")
     ->setFinder(
@@ -42,7 +42,7 @@ return (new PhpCsFixer\Config())
                 'vendor'
             ])
             ->in([
-                __DIR__.'/src/main/php',
+                __DIR__.'/src/app',
             ])
     )
 ;
