@@ -16,7 +16,7 @@
  * @see       https://github.com/rain-noise/sflf/blob/master/src/main/php/Sflf/Form.php
  *
  * @package   SFLF
- * @version   v1.0.2
+ * @version   v1.0.3
  * @author    github.com/rain-noise
  * @copyright Copyright (c) 2017 github.com/rain-noise
  * @license   MIT License https://github.com/rain-noise/sflf/blob/master/LICENSE
@@ -24,7 +24,7 @@
  * @param array{
  *   name?: string|null,
  * }              $params  パラメータ
- * @param \Smarty &$smarty テンプレートオブジェクト
+ * @param Smarty\Smarty &$smarty テンプレートオブジェクト
  * @return mixed|null
  */
 function smarty_function_errors($params, &$smarty)
@@ -59,7 +59,6 @@ function smarty_function_errors($params, &$smarty)
             $html .= '<span class="error">'.$message.'</span>';
         }
     }
-
 
     return $html;
 }
